+++ frontend/src/app/(tabs)/popular.tsx
@@ -2,7 +2,7 @@ import { View } from "react-native";
 
 import { ActivityList } from "@/components/ActivityList";
 import { POPULAR_TITLES } from "@/constants/homeSections";
-import { useActivities } from "@/Context/ActivityContext";
+import { useActivities } from "@/Context/useActivities";
 import { mockActivities } from "@/data/mockActivities";
 import { HomePopUpPageHeaderSection } from "@/home_components/HomePopUpPageHeaderSection";
 import { applyActivityState } from "@/utils/activityState";
